Correct test script in package.json
The test script in the package.json has been updated to call Vitest without the --watch option. Previously, this script was designed to run Vitest in watch mode, but now it will be run independently without monitoring changes. This modification aligns the script execution with other scripts in the package.json file.