Update Square component to accept value prop
The Square function in the App.js file of the Tic Tac Toe React tutorial project has been updated. Previously it was hardcoded to display the value '1'. Now, it accepts a 'value' prop which can be used to control the value it displays. This change allows for dynamic rendering of square values.