Implement winner calculation in tic-tac-toe game
The changes consist in applying a winner calculation function within the handleClick function, to prevent further moves once a player has won. The status of the game (who is the winner, or which player's turn is next) is also displayed on the screen now. Moreover, the winner calculation loop has been changed to a more optimal form.