Refactor App.js to streamline game logic
The variables nextSquare and history have been renamed to nextSquares and history Array initialization respectively for semantic clarity. The if-statement logic inside the handleClick function was simplified, and the Board component now receives xIsNext, currentSquares, and handlePlay as props. This commit concentrates on improving code readability and simplifying board interaction mechanics in 'Tic Tac Toe' game.