Refactor Game function in App.js
Instead of having a "xIsNext" state, we are now using a "currentMove" state in the Game function. This change determines whose turn it is by checking whether "currentMove" is even or odd, and also allows tracking history in a more efficient way. The code for jumping to a specific move was also improved.